+++ src/App/Services.php
@@ -4,7 +4,7 @@ declare(strict_types=1);
 
 use App\Service\Note\NoteService;
 use App\Service\TaskService;
-use App\Service\UserService;
+use App\Service\User\UserService;
 use Psr\Container\ContainerInterface;
 
 $container = $app->getContainer();
